Convert AuthorizationManagementRequest to public interface (#694)
